[WIP] Restore the Lib2P2 builds and implement the latest wire spec
Depends on https://github.com/status-im/nim-eth/pull/54

+++ beacon_chain/conf.nim
@@ -79,7 +79,7 @@ type
 
     of createTestnet:
       networkId* {.
-        desc: "An unique numeric identifier for the network".}: uint64
+        desc: "An unique numeric identifier for the network".}: uint8
 
       validatorsDir* {.
         desc: "Directory containing validator descriptors named vXXXXXXX.deposit.json"
